Make primitive_t a function pointer
This function pointer will now accept the state as well as an
environment to mutate.  As primitives may be called during closure
evaluation, each of which have their own enclosed lexical state, it's
important to pass this to the primitive rather than rely on the global
environment used by the state.